% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3.R
\docType{data}
\name{x}
\alias{x}
\title{A Simple Perceptron Function}
\format{An object of class \code{matrix} with 150 rows and 2 columns.}
\usage{
x
}
\description{
A Simple Perceptron Function
}
\examples{
perceptron()
}
\keyword{perceptron}
